+++ eras/babbage/test-suite/src/Test/Cardano/Ledger/Babbage/TxInfo.hs
@@ -16,7 +16,6 @@ import Cardano.Ledger.Alonzo.Plutus.Context (
   PlutusTxInfo,
  )
 import Cardano.Ledger.Alonzo.Plutus.TxInfo (TxOutSource (..))
-import Cardano.Ledger.Alonzo.Scripts (AlonzoEraScript)
 import Cardano.Ledger.Babbage (BabbageEra)
 import Cardano.Ledger.Babbage.TxBody
 import Cardano.Ledger.Babbage.TxInfo (ContextError (..), transTxInInfoV2, transTxOutV2)
@@ -92,7 +91,7 @@ inlineDatumOutput =
   mkBasicTxOut (shelleyAddr) (inject $ Coin 3)
     & datumTxOutL .~ datumEx
 
-refScriptOutput :: forall era. (BabbageEraTxOut era, AlonzoEraScript era) => TxOut era
+refScriptOutput :: BabbageEraTxOut era => TxOut era
 refScriptOutput =
   mkBasicTxOut (shelleyAddr) (inject $ Coin 3)
     & referenceScriptTxOutL .~ (SJust $ alwaysSucceeds @'PlutusV2 3)
@@ -110,7 +109,6 @@ inputWithRefScript = mkTxInPartial genesisId 4
 
 utxo ::
   ( BabbageEraTxOut era
-  , AlonzoEraScript era
   , Value era ~ MaryValue (EraCrypto era)
   ) =>
   UTxO era
@@ -218,8 +216,7 @@ errorTranslate exampleName =
 
 translatedInputEx1 ::
   forall era.
-  ( AlonzoEraScript era
-  , BabbageEraTxOut era
+  ( BabbageEraTxOut era
   , Show (ContextError era)
   , Value era ~ MaryValue (EraCrypto era)
   , Inject (ContextError (AlonzoEra (EraCrypto era))) (ContextError era)
@@ -232,8 +229,7 @@ translatedInputEx1 _ =
 
 translatedInputEx2 ::
   forall era.
-  ( AlonzoEraScript era
-  , BabbageEraTxOut era
+  ( BabbageEraTxOut era
   , Show (ContextError era)
   , Value era ~ MaryValue (EraCrypto era)
   , Inject (ContextError (AlonzoEra (EraCrypto era))) (ContextError era)
@@ -259,8 +255,7 @@ translatedOutputEx1 _ =
 
 translatedOutputEx2 ::
   forall era.
-  ( AlonzoEraScript era
-  , BabbageEraTxOut era
+  ( BabbageEraTxOut era
   , Show (ContextError era)
   , Value era ~ MaryValue (EraCrypto era)
   , Inject (ContextError (BabbageEra (EraCrypto era))) (ContextError era)
